fshttp: Fix deadlock when RoundTrip() is recursively called
roundTripHook was being called in a lexical scope, but due to how
golang's defer works, roundTripLock.Unlock() was not executed until the
function returned. Since golang's mutexes are not reentrant, this
results in a deadlock.

This is fixed by moving the locking and the roundTripHook() call to a
separate function so that the mutex is unlocked sooner.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>